Median Interview questions
Google interview problems -
My implementation of the coding portion of various interview questions

Median - define a function int median(int a, int b, int c)
that returns the value of the median. The function has 3 arguments.

Median2 - define a function int median(int a, int b, int c, int min, int max)
that returns the value of the median. The function takes 5 arguments.
Min and Max pertain to 2 of the 3 inputs a, b, or c.